Tune compiler flags to more strict

+++ gempyrelib/src/server.h
@@ -33,8 +33,8 @@ public:
     using CloseFunction =  std::function<void (Close, int)>;
     using OpenFunction =  std::function<void (int)>;
     using GetFunction =  std::function<std::optional<std::string> (const std::string_view& filename)>;
-    using ListenFunction =  std::function<bool (unsigned short)>;
-    Server(unsigned short port,
+    using ListenFunction =  std::function<bool (unsigned)>;
+    Server(unsigned int port,
            const std::string& rootFolder,
            const OpenFunction& onOpen,
            const MessageFunction& onMessage,
@@ -51,7 +51,7 @@ public:
     bool send(const std::unordered_map<std::string, std::string>& object, const std::any& values = std::any());
     bool send(const char* data, size_t len);
     int queryId() const {return ++m_queryId;}
-    unsigned short port() const {return m_port;}
+    unsigned int port() const {return m_port;}
     std::function<std::string (const std::string&)> onFile(const std::function<std::string (const std::string&)>&f );
     ~Server();
     bool beginBatch();
@@ -62,11 +62,11 @@ private:
     void closeListenSocket();
     enum class DataType{Json, Bin};
     int addPulled(DataType, const std::string_view& data);
-    void serverThread(unsigned short port);
+    void serverThread(unsigned port);
     bool checkPort();
     std::unique_ptr<std::thread> newThread();
 private:
-    unsigned short m_port;
+    unsigned int m_port;
     std::string m_rootFolder;
     std::unique_ptr<Broadcaster> m_broadcaster;
     std::unique_ptr<Broadcaster> m_extensions;
